Day 10 solution optimization
Rather than store the return string in an array, join it, then
split it for the count, I figured it would be better to concat
each digit to a string, then split that string for the count.
This shaves off a full 10 seconds (25 -> 15) from execution time
on the 50 iteration version.