Day 15 part one answer complete.
I wanted to find a general answer that would work for any number of ingredients in an input, so this took longer than expected. This led down a path of trying to find the fastest combination/permutation module for perl with the winner being ntheory.  Had I tried to solve for only the number of ingredients in the input file, I could've done this with a bunch of for loops.